stacktrace: Record birth time for newly created stack trace
It seems that stack trace that might be leaked usually survive for a
very long time.

Even if some stack trace allocated hugh size and many counts, it's
highly likely to be normally allocations if stack trace is relatively
young so might be deallocated in the near future.

So providing the time when the stack trace wes created can give a hint
to identify which stack trace might be leaked.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>