stacktrace: Add max_count and max_total_size values
This patch introduces maximum values for count and size.  If some stack
trace is unlikely to have a memory leak, it can have some gap between
the value and its max value.

  === stackmap #3 === [count/max: 370/408] [size/max: 38480/42432]
   0 [0x76ecdd44] malloc +0x11
   1 [0x70a7a898] operator new(unsigned int) +0x4
   2 [0x720fecec] libcbe.so (+0x720fecec)
   3 [0x7211dee6] SkPictureRecorder::finishRecordingAsPicture() +0x59
   4 [0x730cb116] blink::GraphicsContext::endRecording() +0x17
   5 [0x730febe0] blink::DrawingRecorder::~DrawingRecorder() +0x9
   6 [0x7415fda8] libcbe.so (+0x7415fda8)
   7 [0x742e0be4] libcbe.so (+0x742e0be4)

In the above output, its count is 370, but the max count is 408.  It
means that the maximum allocation for the same backtrace is done 408
times but its current allocation count is 370 and it represents some of
the allocations are free properly here.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>