libheaptrace: Fix a hanging problem in some hooks
It used to check 'initialized' in a thread local variable, but in some
cases, another thread may jump into the hooked routines after the
initialization.

In this case, it tries to get the function location with dlsym again and
it could make some kind of hanging problems.  It needs to fixed by
checking real_* variables directly.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>