libheaptrace: Fix initialized from TLS to global
The current implementation checks 'initialized' as a TLS variable, but
it's actually incorrect because heaptrace_init() is invoked only once
for each process, not for each thread.

More precisely, it's invoked when exec() invoked to replace its address
map.

Due to this bug, thread allocations are missed and thread deallocations
are also missed so it keeps memory allocation status if the target
backtrace is allocated by a different thread.

This patch makes 'initialized' variable from TLS to global static.

Signed-off-by: Honggyu Kim <hong.gyu.kim@lge.com>