build: Add CMakeLists.txt for cmake build
This patch adds cmake build support and basic usages are as follows.

  $ cmake -B out
  $ make -C out

If you want to change the depth of stacktrace, then DEPTH variable can
be provided as follows.

  $ cmake -B out -DDEPTH=32
  $ make -C out

If DEPTH is not given, then the default depth is 8 as written in
CMakeLists.txt.

We don't remove the existing Make build until we're sure that cmake
looks stable enough.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>